refactor: Move the new ID Generator out of access and into common (#2099)
This will be used by the Basic Reports API.
