+++ src/main/kotlin/org/wfanet/measurement/access/deploy/gcloud/spanner/BUILD.bazel
@@ -16,7 +16,7 @@ kt_jvm_library(
     deps = [
         "//src/main/kotlin/org/wfanet/measurement/access/common:tls_client_principal_mapping",
         "//src/main/kotlin/org/wfanet/measurement/access/deploy/gcloud/spanner/db",
-        "//src/main/kotlin/org/wfanet/measurement/access/service/internal:id_generator",
+        "//src/main/kotlin/org/wfanet/measurement/common:id_generator",
         "//src/main/proto/wfa/measurement/internal/access:principals_service_kt_jvm_grpc_proto",
         "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/gcloud/spanner",
     ],
@@ -39,8 +39,8 @@ kt_jvm_library(
     deps = [
         "//src/main/kotlin/org/wfanet/measurement/access/deploy/gcloud/spanner/db",
         "//src/main/kotlin/org/wfanet/measurement/access/service/internal:errors",
-        "//src/main/kotlin/org/wfanet/measurement/access/service/internal:id_generator",
         "//src/main/kotlin/org/wfanet/measurement/access/service/internal:permission_mapping",
+        "//src/main/kotlin/org/wfanet/measurement/common:id_generator",
         "//src/main/kotlin/org/wfanet/measurement/common/api:etags",
         "//src/main/proto/wfa/measurement/internal/access:roles_service_kt_jvm_grpc_proto",
         "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/gcloud/spanner",
@@ -53,7 +53,7 @@ kt_jvm_library(
     deps = [
         "//src/main/kotlin/org/wfanet/measurement/access/common:tls_client_principal_mapping",
         "//src/main/kotlin/org/wfanet/measurement/access/deploy/gcloud/spanner/db",
-        "//src/main/kotlin/org/wfanet/measurement/access/service/internal:id_generator",
+        "//src/main/kotlin/org/wfanet/measurement/common:id_generator",
         "//src/main/kotlin/org/wfanet/measurement/common/api:etags",
         "//src/main/proto/wfa/measurement/internal/access:policies_service_kt_jvm_grpc_proto",
         "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/gcloud/spanner",
