+++ src/main/kotlin/org/wfanet/measurement/access/service/internal/BUILD.bazel
@@ -21,11 +21,6 @@ kt_jvm_library(
 )
 
 kt_jvm_library(
-    name = "id_generator",
-    srcs = ["IdGenerator.kt"],
-)
-
-kt_jvm_library(
     name = "permission_mapping",
     srcs = ["PermissionMapping.kt"],
     deps = [
