+++ src/main/kotlin/org/wfanet/measurement/access/service/internal/testing/PrincipalsServiceTest.kt
@@ -32,7 +32,7 @@ import org.mockito.kotlin.times
 import org.mockito.kotlin.verify
 import org.wfanet.measurement.access.common.TlsClientPrincipalMapping
 import org.wfanet.measurement.access.service.internal.Errors
-import org.wfanet.measurement.access.service.internal.IdGenerator
+import org.wfanet.measurement.common.IdGenerator
 import org.wfanet.measurement.common.grpc.errorInfo
 import org.wfanet.measurement.common.toInstant
 import org.wfanet.measurement.internal.access.Principal
