+++ src/test/kotlin/org/wfanet/measurement/access/deploy/gcloud/spanner/SpannerPrincipalsServiceTest.kt
@@ -20,8 +20,8 @@ import org.junit.ClassRule
 import org.junit.Rule
 import org.wfanet.measurement.access.common.TlsClientPrincipalMapping
 import org.wfanet.measurement.access.deploy.gcloud.spanner.testing.Schemata
-import org.wfanet.measurement.access.service.internal.IdGenerator
 import org.wfanet.measurement.access.service.internal.testing.PrincipalsServiceTest
+import org.wfanet.measurement.common.IdGenerator
 import org.wfanet.measurement.gcloud.spanner.testing.SpannerEmulatorDatabaseRule
 import org.wfanet.measurement.gcloud.spanner.testing.SpannerEmulatorRule
