+++ src/test/kotlin/org/wfanet/measurement/access/deploy/gcloud/spanner/SpannerRolesServiceTest.kt
@@ -19,9 +19,9 @@ package org.wfanet.measurement.access.deploy.gcloud.spanner
 import org.junit.ClassRule
 import org.junit.Rule
 import org.wfanet.measurement.access.deploy.gcloud.spanner.testing.Schemata
-import org.wfanet.measurement.access.service.internal.IdGenerator
 import org.wfanet.measurement.access.service.internal.PermissionMapping
 import org.wfanet.measurement.access.service.internal.testing.RolesServiceTest
+import org.wfanet.measurement.common.IdGenerator
 import org.wfanet.measurement.gcloud.spanner.testing.SpannerEmulatorDatabaseRule
 import org.wfanet.measurement.gcloud.spanner.testing.SpannerEmulatorRule
